+++ src/Dockerfile
@@ -4,9 +4,8 @@ WORKDIR /src
 
 COPY requirements.txt requirements.txt
 
-RUN --mount=type=cache,target=~/.cache/pip  \
-    pip install -r requirements.txt
-
+# RUN --mount=type=cache,target=~/.cache/pip  pip install -r requirements.txt
+RUN pip install Flask
 COPY . /src
 
 ENV FLASK_APP=app
@@ -14,7 +13,6 @@ ENV FLASK_APP=app
 EXPOSE 5000
 
 WORKDIR /src/api
-
 CMD ["python", "app.py", "--host=0.0.0.0"]
