+++ src/api/app.py
@@ -2,9 +2,9 @@ import logging
 from pathlib import Path
 import os
 
-from src.models.bert.knn_search import KNNSearch
-from src.logs.viberary_logging import ViberaryLogging
-from src.io.redis_conn import RedisConnection
+from models.bert.knn_search import KNNSearch
+from logs.viberary_logging import ViberaryLogging
+from inout.redis_conn import RedisConnection
 
 from flask import Flask
 from flask import render_template, jsonify, request
