--- src/io/redis_conn.py
@@ -1,9 +1,0 @@
-from redis import Redis
-
-
-class RedisConnection:
-    def conn(self) -> Redis:
-        host = "localhost"
-        port = 6379
-        redis_conn = Redis(host=host, port=port, decode_responses=True)
-        return redis_conn
